Changed VotedTalk total ratings to non-floating point value - as in a simple counter

+++ conference-stepengine/src/main/java/org/tweetwallfx/conference/stepengine/dataprovider/VotedTalk.java
@@ -35,7 +35,7 @@ public class VotedTalk {
 
     public final String speakers;
     public final double ratingAverageScore;
-    public final double ratingTotalVotes;
+    public final int ratingTotalVotes;
     public final Speaker speaker;
     public final Talk talk;
